Customers, User, items Controllers fix, SalesController Addition, migration hotfix, and routing fix

+++ database/migrations/2025_04_14_023602_item.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('items', function (Blueprint $table) {
                 $table->uuid('id')->primary();
                 $table->string('name', 255);
-                $table->decimal('price', 8,2);
+                $table->double('price', 8,2);
                 $table->integer('stock', 11);
                 $table->timestamp('created_at');
                 $table->timestamp('updated_at');
